refactor(common): move .keyman-touch-layout reader/writer to @keymanapp/developer-utils
Relates to #9665
