+++ developer/src/common/web/utils/test/fixtures/keyman-touch-layout/legacy.keyman-touch-layout
